refac:UserRegister 수정 및 파일 이름, 위치 수정
