+++ test-web/src/components/Base/UserContext.js
@@ -1,0 +1,24 @@
+// UserContext.js
+
+import { createContext, useContext, useState } from 'react';
+
+const UserContext = createContext();
+
+const UserProvider = ({ children }) => {
+  const [user, setUser] = useState({}); // Your initial user state here
+
+  return (
+    <UserContext.Provider value={{ user, setUser }}>
+      {children}
+    </UserContext.Provider>
+  );
+};
+
+const useUser = () => {
+  const context = useContext(UserContext);
+  if (!context) {
+    throw new Error('useUser must be used within a UserProvider');
+  }
+  return context.user;
+};
+export { UserProvider, useUser };
