+++ test-web/src/components/Charts/PieChart/helper/usePieChart.js
@@ -2,8 +2,6 @@ import merge from 'lodash/merge';
 // @mui
 import { useTheme, alpha } from '@mui/material/styles';
 
-// ----------------------------------------------------------------------
-
 export default function useChart(options) {
   const theme = useTheme();
